👷 Move entity to pkg
To simplify creation of integration tests.
