+++ internal/usecase/usecase.go
@@ -5,8 +5,8 @@ import (
 	"time"
 
 	"github.com/VetKA-org/vetka/internal/config"
-	"github.com/VetKA-org/vetka/internal/entity"
 	"github.com/VetKA-org/vetka/internal/repo"
+	"github.com/VetKA-org/vetka/pkg/entity"
 	"github.com/VetKA-org/vetka/pkg/logger"
 	uuid "github.com/satori/go.uuid"
 )
