+++ pkg/entity/appointment.go
