+++ pkg/entity/role.go
