+++ pkg/entity/secret.go
