+++ pkg/entity/species.go
